+++ app/controllers/users_controller.rb
@@ -53,11 +53,11 @@ class UsersController < ApplicationController
 
       if current_user.system_admin?
         params.require(:user).permit(
-          :first_name, :last_name, :email, :username, :email_me_when_mentioned, :research_summary, :degree, :aug_member, :core_member, :system_admin, :banned, :auto_subscribe
+          :first_name, :last_name, :email, :username, :research_summary, :degree, :aug_member, :core_member, :system_admin, :banned, :emails_enabled
         )
       else
         params.require(:user).permit(
-          :first_name, :last_name, :email, :username, :email_me_when_mentioned, :research_summary, :auto_subscribe
+          :first_name, :last_name, :email, :username, :research_summary, :emails_enabled
         )
       end
     end