+++ app/models/post.rb
@@ -1,0 +1,2 @@
+class Post < Comment  # ActiveRecord::Base
+end