+++ db/migrate/20150604131345_change_auto_subscribe_for_users.rb
@@ -1,0 +1,9 @@
+class ChangeAutoSubscribeForUsers < ActiveRecord::Migration
+  def up
+    change_column :users, :auto_subscribe, :boolean, null: false, default: false
+  end
+
+  def down
+    change_column :users, :auto_subscribe, :boolean, null: false, default: true
+  end
+end