+++ db/migrate/20150604140623_remove_forum_email_flags_from_users.rb
@@ -1,0 +1,6 @@
+class RemoveForumEmailFlagsFromUsers < ActiveRecord::Migration
+  def change
+    remove_column :users, :auto_subscribe, null: false, default: false
+    remove_column :users, :email_me_when_mentioned, null: false, default: true
+  end
+end